2023.08.09

+++ src/main/java/com/matsinger/barofishserver/grade/api/GradeController.java
@@ -57,7 +57,7 @@ public class GradeController {
     @NoArgsConstructor
     private static class AddGradeReq {
         String name;
-        Integer pointRate;
+        Float pointRate;
         Integer minOrderPrice;
         Integer minOrderCount;
     }
